separate plotting and training
