+++ src/spaceoracle/oracles.py
@@ -394,125 +394,6 @@ class SpaceTravLR(BaseTravLR):
             train_bar.count = 0
             train_bar.start = time.time()
 
-
-    @staticmethod
-    def load_betadata(gene, save_dir):
-        return pd.read_parquet(f'{save_dir}/{gene}_betadata.parquet')
-    
-    
-    def _get_wbetas_dict(self, betas_dict, gene_mtx):
-        
-        gex_df = pd.DataFrame(gene_mtx, index=self.adata.obs_names, columns=self.adata.var_names)
-
-        if len(self.ligands) > 0:
-            weighted_ligands = received_ligands(
-                xy=self.adata.obsm['spatial'], 
-                lig_df=gex_df[list(self.ligands)],
-                radius=self.radius
-            )
-        else:
-            weighted_ligands = []
-
-        self.weighted_ligands = weighted_ligands
-
-        for gene, betadata in tqdm(betas_dict.data.items(), total=len(betas_dict), desc='Interactions', disable=len(betas_dict) == 1):
-            betas_dict.data[gene].wbetas = self._combine_gene_wbetas(gene, weighted_ligands, gex_df, betadata)
-
-        # for gene, betaoutput in tqdm(betas_dict.items(), total=len(betas_dict), desc='Ligand interactions', disable=len(betas_dict) == 1):
-        #     betas_df = self._combine_gene_wbetas(gene, weighted_ligands, gex_df, betaoutput)
-        #     betas_dict[gene].wbetas = betas_df
-
-        return betas_dict
-
-    def _combine_gene_wbetas(self, gene, rw_ligands, gex_df, betadata):
-        betas_df = betadata.splash(rw_ligands, gex_df)
-        return betas_df
-        
-
-    def _get_spatial_betas_dict(self):
-        bdb = Betabase(self.adata, self.save_dir)
-        self.ligands = bdb.ligands_set
-        return bdb
-    
-
-    def _perturb_single_cell(self, gex_delta, cell_index, betas_dict):
-
-        genes = self.adata.var_names
-        
-        gene_gene_matrix = np.zeros((len(genes), len(genes))) # columns are target genes, rows are regulators
-
-        for i, gene in enumerate(genes):
-            _beta_out = betas_dict.data.get(gene, None)
-            
-            if _beta_out is not None:
-                r = np.array(_beta_out.modulator_gene_indices)
-                gene_gene_matrix[r, i] = _beta_out.wbetas.values[cell_index]
-
-        return gex_delta[cell_index, :].dot(gene_gene_matrix)
-
-
-    def perturb(self, target, gene_mtx=None, n_propagation=3, gene_expr=0, cells=None):
-        
-        for key in ['transition_probabilities', 'grid_points', 'vector_field']:
-            self.adata.uns.pop(key, None)
-
-        assert target in self.adata.var_names
-        
-        if gene_mtx is None: 
-            gene_mtx = self.adata.layers['imputed_count']
-
-        if isinstance(gene_mtx, pd.DataFrame):
-            gene_mtx = gene_mtx.values
-
-
-        target_index = self.gene2index[target]  
-        simulation_input = gene_mtx.copy()
-
-        # perturb target gene
-        if cells is None:
-            simulation_input[:, target_index] = gene_expr   
-        else:
-            simulation_input[cells, target_index] = gene_expr
-        
-        delta_input = simulation_input - gene_mtx       # get delta X
-        delta_simulated = delta_input.copy() 
-
-        if self.beta_dict is None:
-            print('Computing beta_dict')
-            self.beta_dict = self._get_spatial_betas_dict() # compute betas for all genes for all cells
-
-        for n in range(n_propagation):
-
-            beta_dict = self._get_wbetas_dict(self.beta_dict, gene_mtx + delta_simulated)
-
-            _simulated = np.array(
-                [self._perturb_single_cell(delta_simulated, i, beta_dict) 
-                    for i in tqdm(
-                        range(self.adata.n_obs), 
-                        desc=f'Running simulation {n+1}/{n_propagation}')])
-            delta_simulated = np.array(_simulated)
-            delta_simulated = np.where(delta_input != 0, delta_input, delta_simulated)
-
-            gem_tmp = gene_mtx + delta_simulated
-            gem_tmp[gem_tmp<0] = 0
-            delta_simulated = gem_tmp - gene_mtx
-
-        gem_simulated = gene_mtx + delta_simulated
-        
-        assert gem_simulated.shape == gene_mtx.shape
-
-        # just as in CellOracle, don't allow simulated to exceed observed values
-        imputed_count = gene_mtx
-        min_ = imputed_count.min(axis=0)
-        max_ = imputed_count.max(axis=0)
-        gem_simulated = pd.DataFrame(gem_simulated).clip(lower=min_, upper=max_, axis=1).values
-
-        self.adata.layers['simulated_count'] = gem_simulated
-        self.adata.layers['delta_X'] = gem_simulated - imputed_count
-
-        return gem_simulated
-
-
     @staticmethod
     def imbue_adata_with_space(adata, annot='rctd_cluster', spatial_dim=64, in_place=False, method='fast'):
         clusters = np.array(adata.obs[annot])
@@ -546,12 +427,8 @@ class SpaceTravLR(BaseTravLR):
 
         return sp_maps
 
-
-    def compute_betas(self):
-        self.beta_dict = self._get_spatial_betas_dict()
-        self.coef_matrix = self._get_co_betas()
-
-
+    # This is incorrect, needs to be cluster specific
+    # Load links from CO trained model
     def _get_co_betas(self, alpha=1):
 
         gem = self.adata.to_df(layer='imputed_count')
